src/third_party/blink/web_tests/wpt_internal/display-lock/css-render-subtree/activation/accessibility-activatable-press.html

<!DOCTYPE HTML>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>

<div id="locked" style="render-subtree: invisible">
  foo
  <div id="child" tabindex="0">
    bar
  </div>
</div>

<script>
function axElementById(id) {
  return accessibilityController.accessibleElementById(id);
}

async_test(async(t) => {
  const lockedEl = document.getElementById("locked");
  let axLocked = axElementById("locked");
  t.step(() => { assert_equals(axLocked.childrenCount, 3, "Child count after acquire"); });

  axElementById("child").press();

  // Wait for the next frame for the ax object to be recreated.
  requestAnimationFrame(() => {
    requestAnimationFrame(() => {
      axLocked = axElementById("locked");
      t.step(() => { assert_equals(axLocked.childrenCount, 2, "Child count after activation"); });
      t.done();
    });
  });
}, "Accessiblility press() causes activatable locked tree to activate");

</script>
